refactor(common): place the thermodynamics operators and docstrings as reviewed
Review round on the regrouped package, applied as proposed:

- `compute_temperature_from_internal_energy` and its scalar variant move from
  `compute_energy.py` to `compute_temperature.py`, since they compute a
  temperature; their stencil test follows
- "internal energy" is an internal energy per unit area (the operators
  multiply by rho * dz), and every docstring now says so
- `dz` documented as the vertical extent of the grid cell
- both module docstrings dropped; the surface-pressure extrapolation sentence
  moves onto its stencil, the hydrostatic-pressure stencil gets the fuller
  docstring from the review

From the pre-push review: the hydrostatic balance in that docstring read
`dp/dz = -rho g = p g / (Rd Tv)` -- the minus was missing on the right-hand
side -- fixed and wrapped; `c_v * rho * dz` is a heat capacity per unit area,
not a specific heat; and the note that constants must come from a wpfloat
Enum for gtfn now sits on `PhysicsConstants` itself rather than at an import.
