+++ model/common/src/icon4py/model/common/physics/thermodynamics/compute_energy.py
@@ -5,97 +5,11 @@
 #
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
-"""
-Moist internal energy and its inverse, the temperature.
-
-The internal energy of a grid cell follows from the specific heats of dry air,
-vapour, liquid and ice weighted by their mass fractions, minus the latent heats
-of the condensed phases; the inverse relation recovers the temperature from it.
-
-The constants come from :class:`icon4py.model.common.constants.PhysicsConstants`
-(a ``wpfloat``-based ``enum.Enum``, which is what GT4Py/gtfn needs for symbols
-referenced inside field operators).
-"""
-
 import gt4py.next as gtx
 
 from icon4py.model.common import field_type_aliases as fa, type_alias as ta
 from icon4py.model.common.constants import PhysicsConstants
 from icon4py.model.common.type_alias import wpfloat
-
-
-@gtx.field_operator(grid_type=gtx.GridType.UNSTRUCTURED)
-def compute_temperature_from_internal_energy(  # noqa: PLR0917 [too-many-positional-arguments]
-    u: fa.CellKField[ta.wpfloat],
-    qv: fa.CellKField[ta.wpfloat],
-    qliq: fa.CellKField[ta.wpfloat],
-    qice: fa.CellKField[ta.wpfloat],
-    rho: fa.CellKField[ta.wpfloat],
-    dz: fa.CellKField[ta.wpfloat],
-) -> fa.CellKField[ta.wpfloat]:
-    """
-    Compute the temperature from the internal energy
-
-    Args:
-        u:                  Internal energy (extensive)
-        qv:                 Water vapor specific humidity
-        qliq:               Specific mass of liquid phases
-        qice:               Specific mass of solid phases
-        rho:                Ambient density
-        dz:                 Extent of grid cell
-
-    Return:                 Temperature
-    """
-    qtot = qliq + qice + qv  # total water specific mass
-    cv = (
-        (
-            PhysicsConstants.cvd * (wpfloat(1.0) - qtot)
-            + PhysicsConstants.cvv * qv
-            + PhysicsConstants.cpl * qliq
-            + PhysicsConstants.cpi * qice
-        )
-        * rho
-        * dz
-    )  # Moist isometric specific heat
-
-    return (u + rho * dz * (qliq * PhysicsConstants.lvc + qice * PhysicsConstants.lsc)) / cv
-
-
-@gtx.field_operator
-def compute_temperature_from_internal_energy_scalar(  # noqa: PLR0917 [too-many-positional-arguments]
-    u: ta.wpfloat,
-    qv: ta.wpfloat,
-    qliq: ta.wpfloat,
-    qice: ta.wpfloat,
-    rho: ta.wpfloat,
-    dz: ta.wpfloat,
-) -> ta.wpfloat:
-    """
-    Compute the temperature from the internal energy (scalar version callable from scan_operator)
-
-    Args:
-        u:                  Internal energy (extensive)
-        qv:                 Water vapor specific humidity
-        qliq:               Specific mass of liquid phases
-        qice:               Specific mass of solid phases
-        rho:                Ambient density
-        dz:                 Extent of grid cell
-
-    Return:                 Temperature
-    """
-    qtot = qliq + qice + qv  # total water specific mass
-    cv = (
-        (
-            PhysicsConstants.cvd * (wpfloat(1.0) - qtot)
-            + PhysicsConstants.cvv * qv
-            + PhysicsConstants.cpl * qliq
-            + PhysicsConstants.cpi * qice
-        )
-        * rho
-        * dz
-    )  # Moist isometric specific heat
-
-    return (u + rho * dz * (qliq * PhysicsConstants.lvc + qice * PhysicsConstants.lsc)) / cv
 
 
 @gtx.field_operator(grid_type=gtx.GridType.UNSTRUCTURED)
@@ -108,7 +22,7 @@ def compute_internal_energy(  # noqa: PLR0917 [too-many-positional-arguments]
     dz: fa.CellKField[ta.wpfloat],
 ) -> fa.CellKField[ta.wpfloat]:
     """
-    Compute the internal energy from the temperature
+    Compute the internal energy per unit area from the temperature
 
     Args:
         t:                 Temperature
@@ -116,9 +30,9 @@ def compute_internal_energy(  # noqa: PLR0917 [too-many-positional-arguments]
         qliq:              Specific mass of liquid phases
         qice:              Specific mass of solid phases
         rho:               Ambient density
-        dz:                Extent of grid cell
+        dz:                Vertical extent of grid cell
 
-    Result:                Internal energy
+    Result:                Internal energy per unit area
     """
     qtot = qliq + qice + qv
     cv = (
@@ -141,7 +55,7 @@ def compute_internal_energy_scalar(  # noqa: PLR0917 [too-many-positional-argume
     dz: ta.wpfloat,
 ) -> ta.wpfloat:
     """
-    Compute the internal energy from the temperature
+    Compute the internal energy per unit area from the temperature
 
     Args:
         t:                 Temperature
@@ -149,9 +63,9 @@ def compute_internal_energy_scalar(  # noqa: PLR0917 [too-many-positional-argume
         qliq:              Specific mass of liquid phases
         qice:              Specific mass of solid phases
         rho:               Ambient density
-        dz:                Extent of grid cell
+        dz:                Vertical extent of grid cell
 
-    Result:                Internal energy
+    Result:                Internal energy per unit area
     """
     qtot = qliq + qice + qv
     cv = (
