+++ model/common/tests/common/physics/stencil_tests/test_compute_temperature.py
@@ -14,6 +14,7 @@ import pytest
 from icon4py.model.common import constants as phy_const, dimension as dims, type_alias as ta
 from icon4py.model.common.grid import base
 from icon4py.model.common.physics.thermodynamics.compute_temperature import (
+    compute_temperature_from_internal_energy,
     compute_virtual_temperature_and_temperature,
 )
 from icon4py.model.testing import stencil_tests
@@ -77,4 +78,41 @@ class TestComputeVirtualTemperatureAndTemperature(stencil_tests.StencilTest):
             horizontal_end=gtx.int32(grid.num_cells),
             vertical_start=gtx.int32(0),
             vertical_end=gtx.int32(grid.num_levels),
+        )
+
+
+class TestComputeTemperatureFromInternalEnergy(stencil_tests.StencilTest):
+    PROGRAM = compute_temperature_from_internal_energy
+    OUTPUTS = ("out",)
+
+    @stencil_tests.static_reference
+    def reference(
+        grid: base.Grid,
+        *,
+        u: np.ndarray,
+        qv: np.ndarray,
+        qliq: np.ndarray,
+        qice: np.ndarray,
+        rho: np.ndarray,
+        dz: np.ndarray,
+        **kwargs,
+    ) -> dict:
+        return dict(out=np.full(u.shape, 255.75599999999997))
+
+    @stencil_tests.input_data_fixture
+    def input_data(data_alloc: stencil_tests.DataAllocationWrapper, grid: base.Grid):
+        return dict(
+            u=data_alloc.constant_field(
+                38265357.270336017, dims.CellDim, dims.KDim, dtype=ta.wpfloat
+            ),
+            qv=data_alloc.constant_field(0.00122576, dims.CellDim, dims.KDim, dtype=ta.wpfloat),
+            qliq=data_alloc.constant_field(1.63837e-20, dims.CellDim, dims.KDim, dtype=ta.wpfloat),
+            qice=data_alloc.constant_field(1.09462e-08, dims.CellDim, dims.KDim, dtype=ta.wpfloat),
+            rho=data_alloc.constant_field(0.83444, dims.CellDim, dims.KDim, dtype=ta.wpfloat),
+            dz=data_alloc.constant_field(249.569, dims.CellDim, dims.KDim, dtype=ta.wpfloat),
+            domain={
+                dims.CellDim: (0, gtx.int32(grid.num_cells)),
+                dims.KDim: (0, gtx.int32(grid.num_levels)),
+            },
+            out=data_alloc.zero_field(dims.CellDim, dims.KDim, dtype=ta.wpfloat),
         )
